fix(rtc): mark participants disconnected when the room disconnects
`Participant.state` reads `ParticipantInfo.state`, which only ever moved on
per-participant events. A room-level disconnect — explicit `disconnect()` or an
FFI-driven one — is not reported as each participant departing, so the local
participant and every still-listed remote participant kept reporting ACTIVE.

That state outlives the disconnect in practice: the participant maps are never
cleared, and `Disconnected` handlers routinely capture participants. A retained
participant claiming to be ACTIVE reads as reachable when it no longer is.

Transition the local participant and all retained remote participants to
DISCONNECTED in cleanupOnDisconnect, before `ConnectionStateChanged` and
`Disconnected` fire, so a handler that inspects `state` sees the truth.

Two remote-participant stubs in audio_stream_room_lifecycle.test.ts gain the
`info` that real participants always carry (createRemoteParticipant sets it
unconditionally), since cleanup now writes through it.

Renames participant_active.test.ts to participant_state.test.ts — it now covers
the full state lifecycle rather than just the active transition.
