+++ .changeset/participant-active-event.md
@@ -8,3 +8,7 @@ A remote participant can only receive data messages once it reaches `Participant
 until now JS had no way to observe that transition, so code waiting on `ParticipantConnected`
 could send to a participant that was not yet reachable. This brings the Node SDK in line with the
 Python SDK's `participant_active` event.
+
+`Participant.state` also now reports `DISCONNECTED` once the participant is gone — both when it
+departs individually and when the room itself disconnects, since a room-level disconnect is not
+reported as each participant departing.
